Moved all schemas to single file
