--- schemas/documentation_generation_enums.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-class DocsStatusEnum(str, Enum):
-    STARTED = 'STARTED'
-    COMPLETED = 'COMPLETED'
