--- schemas/generated_doc.py
@@ -1,4 +1,0 @@
-from pydantic import BaseModel
-
-class GeneratedDocResponse(BaseModel):
-    content: str
